Refactor options. Update deps
