+++ v3/pkg/application/application_options.go
@@ -114,3 +114,56 @@ func ChainMiddleware(middleware ...Middleware) Middleware {
 func AssetFileServerFS(assets fs.FS) http.Handler {
 	return assetserver.NewAssetFileServer(assets)
 }
+
+/******** Mac Options ********/
+
+// ActivationPolicy is the activation policy for the application.
+type ActivationPolicy int
+
+const (
+	// ActivationPolicyRegular is used for applications that have a user interface,
+	ActivationPolicyRegular ActivationPolicy = iota
+	// ActivationPolicyAccessory is used for applications that do not have a main window,
+	// such as system tray applications or background applications.
+	ActivationPolicyAccessory
+	ActivationPolicyProhibited
+)
+
+// MacOptions contains options for macOS applications.
+type MacOptions struct {
+	// ActivationPolicy is the activation policy for the application. Defaults to
+	// applicationActivationPolicyRegular.
+	ActivationPolicy ActivationPolicy
+	// If set to true, the application will terminate when the last window is closed.
+	ApplicationShouldTerminateAfterLastWindowClosed bool
+}
+
+/****** Windows Options *******/
+
+// WindowsOptions contains options for Windows applications.
+type WindowsOptions struct {
+
+	// WndProcInterceptor is a function that will be called for every message sent in the application.
+	// Use this to hook into the main message loop. This is useful for handling custom window messages.
+	// If `shouldReturn` is `true` then `returnCode` will be returned by the main message loop.
+	// If `shouldReturn` is `false` then returnCode will be ignored and the message will be processed by the main message loop.
+	WndProcInterceptor func(hwnd uintptr, msg uint32, wParam, lParam uintptr) (returnCode uintptr, shouldReturn bool)
+
+	// DisableQuitOnLastWindowClosed disables the auto quit of the application if the last window has been closed.
+	DisableQuitOnLastWindowClosed bool
+
+	// Path where the WebView2 stores the user data. If empty %APPDATA%\[BinaryName.exe] will be used.
+	// If the path is not valid, a messagebox will be displayed with the error and the app will exit with error code.
+	WebviewUserDataPath string
+
+	// Path to the directory with WebView2 executables. If empty WebView2 installed in the system will be used.
+	WebviewBrowserPath string
+}
+
+/********* Linux Options *********/
+
+// LinuxOptions contains options for Linux applications.
+type LinuxOptions struct {
+	// DisableQuitOnLastWindowClosed disables the auto quit of the application if the last window has been closed.
+	DisableQuitOnLastWindowClosed bool
+}
